Lots of clean up
