+++ src/config.ts
@@ -1,5 +1,6 @@
 
 export const config = {
+  // TODO... We might want to palce the store ID here
   clientId: 'EdP3Gi1agyUF3yFS7Ngm8iyodLgbSR3wY4ceoJl0d2',
   stripeKey: 'pk_test_JwRX4cevuCysIEQrpwcbn3j8',
   categoryPageSize: 8,
